Refactoring(server-info): moved main logic from cli to Server class.
