--- netutils_linux_hardware/collector.py
@@ -1,11 +1,0 @@
-# coding=utf-8
-
-
-class Collector(object):
-    @staticmethod
-    def copyfile(src, dst):
-        pass
-
-    @staticmethod
-    def catch_stdout(cmd, dst):
-        pass
